Change Variable flags to a byte

+++ KirbyLib/Mint/MintVariable.cs
@@ -23,7 +23,7 @@ namespace KirbyLib.Mint
         /// <summary>
         /// Bitflags that tell the Mint VM how to treat this variable.
         /// </summary>
-        public uint Flags;
+        public byte Flags;
 
         public MintVariable(string type, string name)
         {
